Added Main Code into Repository for Submission
This repository contains the full implementation of the SDSS Galaxy Clustering and BAO Peak Detection Analysis project. It includes Source Code (galaxycorr), and Test Suites (tests/tests_galaxycorr). The data of SDSS DR17 galaxies used for this project is listed in (galaxycoor\data\sdss_galaxies.fits). This package is designed to be fully reproducible and easily extendable to larger SDSS datasets or simulated galaxy catalogs.